Move the variance app and calculator into their own directory.
