+++ Variance/varianceCalculator.py
